requirements/base.txt
fastapi==0.87.0
SQLAlchemy==1.4.44
uvicorn==0.20.0
ffprobe==0.5
pydub==0.25.1
SpeechRecognition==3.9.0
transformers==4.25.1
fasteners==0.18
Jinja2==3.1.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
databases==0.6.1
aiosqlite==0.17.0
fastapi-limiter==0.1.5
aioredis==2.0.1
loguru==0.6.0
sqlalchemy-file==0.2.0
tensorflow==2.11.0
pydantic[email]
passlib[bcrypt]
databases[sqlite]
fastapi[all]
